Move circular_velocity to new module
